requirements.txt
tensorflow-gpu
keras
matplotlib
numpy
cv2
tqdm
carla